msi: refactoring to a versioned folder
