+++ internal/services/msi/sdk/2018-11-30/managedidentity/id_userassignedidentities.go
