+++ internal/services/msi/user_assigned_identity_resource.go
@@ -5,13 +5,14 @@ import (
 	"log"
 	"time"
 
+	managedidentity2 "github.com/hashicorp/terraform-provider-azurerm/internal/services/msi/sdk/2018-11-30/managedidentity"
+
 	"github.com/hashicorp/go-azure-helpers/response"
 	"github.com/hashicorp/terraform-provider-azurerm/helpers/azure"
 	"github.com/hashicorp/terraform-provider-azurerm/helpers/tf"
 	"github.com/hashicorp/terraform-provider-azurerm/internal/clients"
 	"github.com/hashicorp/terraform-provider-azurerm/internal/location"
 	"github.com/hashicorp/terraform-provider-azurerm/internal/services/msi/migration"
-	"github.com/hashicorp/terraform-provider-azurerm/internal/services/msi/sdk/managedidentity"
 	"github.com/hashicorp/terraform-provider-azurerm/internal/tags"
 	"github.com/hashicorp/terraform-provider-azurerm/internal/tf/pluginsdk"
 	"github.com/hashicorp/terraform-provider-azurerm/internal/tf/validation"
@@ -26,7 +27,7 @@ func resourceArmUserAssignedIdentity() *pluginsdk.Resource {
 		Update: resourceArmUserAssignedIdentityCreateUpdate,
 		Delete: resourceArmUserAssignedIdentityDelete,
 		Importer: pluginsdk.ImporterValidatingResourceId(func(id string) error {
-			_, err := managedidentity.ParseUserAssignedIdentitiesID(id)
+			_, err := managedidentity2.ParseUserAssignedIdentitiesID(id)
 			return err
 		}),
 
@@ -85,7 +86,7 @@ func resourceArmUserAssignedIdentityCreateUpdate(d *pluginsdk.ResourceData, meta
 	location := d.Get("location").(string)
 	t := d.Get("tags").(map[string]interface{})
 
-	resourceId := managedidentity.NewUserAssignedIdentitiesID(subscriptionId, d.Get("resource_group_name").(string), d.Get("name").(string))
+	resourceId := managedidentity2.NewUserAssignedIdentitiesID(subscriptionId, d.Get("resource_group_name").(string), d.Get("name").(string))
 	if d.IsNewResource() {
 		existing, err := client.UserAssignedIdentitiesGet(ctx, resourceId)
 		if err != nil {
@@ -99,7 +100,7 @@ func resourceArmUserAssignedIdentityCreateUpdate(d *pluginsdk.ResourceData, meta
 		}
 	}
 
-	identity := managedidentity.Identity{
+	identity := managedidentity2.Identity{
 		Name:     utils.String(resourceId.UserAssignedIdentityName),
 		Location: location,
 		Tags:     expandTags(t),
@@ -118,7 +119,7 @@ func resourceArmUserAssignedIdentityRead(d *pluginsdk.ResourceData, meta interfa
 	ctx, cancel := timeouts.ForRead(meta.(*clients.Client).StopContext, d)
 	defer cancel()
 
-	id, err := managedidentity.ParseUserAssignedIdentitiesID(d.Id())
+	id, err := managedidentity2.ParseUserAssignedIdentitiesID(d.Id())
 	if err != nil {
 		return err
 	}
@@ -157,7 +158,7 @@ func resourceArmUserAssignedIdentityDelete(d *pluginsdk.ResourceData, meta inter
 	ctx, cancel := timeouts.ForDelete(meta.(*clients.Client).StopContext, d)
 	defer cancel()
 
-	id, err := managedidentity.ParseUserAssignedIdentitiesID(d.Id())
+	id, err := managedidentity2.ParseUserAssignedIdentitiesID(d.Id())
 	if err != nil {
 		return err
 	}
