+++ internal/services/msi/validate/user_assigned_identity_id.go
@@ -3,7 +3,7 @@ package validate
 import (
 	"fmt"
 
-	"github.com/hashicorp/terraform-provider-azurerm/internal/services/msi/sdk/managedidentity"
+	"github.com/hashicorp/terraform-provider-azurerm/internal/services/msi/sdk/2018-11-30/managedidentity"
 )
 
 func UserAssignedIdentityID(input interface{}, key string) (warnings []string, errors []error) {
